// src/app/types/dice-types.ts
export enum DiceType {
  D4 = 'd4',
  D6 = 'd6',
  D8 = 'd8',
  D10 = 'd10',
  D12 = 'd12',
  D20 = 'd20',
  D100 = 'd100'
}

export enum AdvantageType {
  NONE = 'none',
  ADVANTAGE = 'advantage',
  DISADVANTAGE = 'disadvantage'
}

export enum KeepDropType {
  KEEP_HIGHEST = 'keep_highest',
  KEEP_LOWEST = 'keep_lowest',
  DROP_HIGHEST = 'drop_highest',
  DROP_LOWEST = 'drop_lowest'
}